requirements.txt
beautifulsoup4==4.11.1
boto3==1.24.3
pandas==1.4.2
PyYAML==6.0
selenium==4.5.0
selenium_wire==4.6.5
Unidecode==1.3.6
pyOpenSSL==22.0.0
webdriver-manager